Moved cast module to playback folder and added :playback:base module
