+++ app/build.gradle
@@ -112,10 +112,10 @@ android {
 dependencies {
     implementation project(":core")
     implementation project(':model')
-    implementation project(':net:cast')
     implementation project(':net:sync:gpoddernet')
     implementation project(':net:sync:model')
     implementation project(':parser:feed')
+    implementation project(':playback:cast')
     implementation project(':ui:app-start-intent')
     implementation project(':ui:common')
