+++ core/build.gradle
@@ -21,12 +21,12 @@ android {
 
 dependencies {
     implementation project(':model')
-    implementation project(':net:cast')
     implementation project(':net:ssl')
     implementation project(':net:sync:gpoddernet')
     implementation project(':net:sync:model')
     implementation project(':parser:feed')
     implementation project(':parser:media')
+    implementation project(':playback:cast')
     implementation project(':ui:app-start-intent')
     implementation project(':ui:common')
     implementation project(':ui:png-icons')
