--- net/cast/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="de.danoeh.antennapod.net.cast" />
