+++ playback/README.md
@@ -1,0 +1,3 @@
+# :playback
+
+This folder contains modules that deal with media playback.
