+++ playback/cast/README.md
@@ -1,3 +1,3 @@
-# :net:cast
+# :playback:cast
 
 This module provides Chromecast support for the Google Play version of the app.
