+++ settings.gradle
@@ -2,13 +2,15 @@ include ':app'
 include ':core'
 include ':model'
 
-include ':net:cast'
 include ':net:ssl'
 include ':net:sync:gpoddernet'
 include ':net:sync:model'
 
 include ':parser:feed'
 include ':parser:media'
+
+include ':playback:base'
+include ':playback:cast'
 
 include ':ui:app-start-intent'
 include ':ui:common'
